Move old Gauss PSF code into psf_gauss.py
