+++ gammapy/irf/__init__.py
@@ -4,13 +4,12 @@ Instrument response functions (IRFs).
 """
 from .effective_area import *
 from .background import *
-from .psf_core import *
+from .energy_dispersion import *
 from .psf_table import *
 from .psf_3d import *
-from .psf_check import *
 from .psf_gauss import *
 from .psf_king import *
-from .energy_dispersion import *
+from .psf_check import *
 from .irf_stack import *
 from .io import *
